requirements_dev.txt
fhirpath>=0.4.0
guillotina>=5.0.15
guillotina-elasticsearch>=5.0.0
aioelasticsearch==0.5.2
aiohttp==3.5.4
alabaster==0.7.12
appdirs==1.4.3
argon2-cffi==19.1.0
asn1crypto==1.0.1
astunparse==1.6.2
async-timeout==3.0.1
asyncpg==0.18.3
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backoff==1.8.0
black==19.3b0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
check-manifest==0.39
Click==7.0
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
docker==4.1.0
docutils==0.15.2
elasticsearch==6.4.0
entrypoints==0.3
fhir.resources==5.0.1
filelock==3.0.12
flake8==3.7.8
flake8-isort==2.7.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
isodate==0.6.0
isort==4.3.21
Jinja2==2.10.3
jsonpatch==1.24
jsonpointer==2.0
jsonschema==3.0.2
jwcrypto==0.6.0
lru-dict==1.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
multidict==4.5.2
mypy-extensions==0.4.2
packaging==19.2
pkginfo==1.5.0.1
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.0
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pyroma==2.6b1
pyrsistent==0.15.4
pytest==5.2.1
pytest-docker-fixtures==1.3.5
pytest-mock==1.11.1
pytest-runner==5.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
snowballstemmer==2.0.0
Sphinx==1.8.5
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-httpexample==0.10.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
testfixtures==6.10.0
toml==0.10.0
tox==3.14.0
tqdm==4.36.1
twine==2.0.0
ujson==1.35
urllib3==1.25.6
virtualenv==16.7.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
yarl==1.4.0a0
zc.buildout==2.13.1
zest.releaser==6.19.0
zipp==0.6.0
zope.component==4.5
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.proxy==4.3.2
